separete web from service interfaces
