+++ services/web/server/src/simcore_service_webserver/login/_controller/rest/change.py
@@ -15,7 +15,7 @@ from simcore_service_webserver.db.plugin import get_asyncpg_engine
 from ...._meta import API_VTAG
 from ....products import products_web
 from ....products.models import Product
-from ....security import security_web
+from ....security import security_service
 from ....users import api as users_service
 from ....utils import HOUR
 from ....utils_rate_limiting import global_rate_limit_route
@@ -293,7 +293,7 @@ async def change_password(request: web.Request):
     user = await db.get_user({"id": request[RQT_USERID_KEY]})
     assert user  # nosec
 
-    if not security_web.check_password(
+    if not security_service.check_password(
         passwords.current.get_secret_value(), user["password_hash"]
     ):
         raise web.HTTPUnprocessableEntity(
@@ -303,7 +303,7 @@ async def change_password(request: web.Request):
     await db.update_user(
         dict(user),
         {
-            "password_hash": security_web.encrypt_password(
+            "password_hash": security_service.encrypt_password(
                 passwords.new.get_secret_value()
             )
         },
