+++ services/web/server/src/simcore_service_webserver/security/security_service.py
@@ -1,0 +1,20 @@
+"""Service-layer interface
+
+NOTE: Must not include functions that depend on aiohttp.web.Request (use *_web.py instead)
+"""
+
+from ._authz_service import (
+    check_password,
+    clean_auth_policy_cache,
+    encrypt_password,
+    get_access_model,
+)
+
+__all__: tuple[str, ...] = (
+    "check_password",
+    "clean_auth_policy_cache",
+    "encrypt_password",
+    "get_access_model",
+)
+
+# nopycln: file
