+++ services/web/server/src/simcore_service_webserver/security/security_web.py
@@ -1,17 +1,14 @@
 # mypy: disable-error-code=truthy-function
-"""
+"""aiohttp.web-related interfaces i.e. web.Request is used in the inputs
 
 NOTE: DO NOT USE aiohttp_security.api directly but use this interface instead
+NOTE: functions in this module
 """
 
 from ._authz_access_model import AuthContextDict
-from ._authz_service import (
-    check_password,
+from ._authz_web import (
     check_user_authorized,
     check_user_permission,
-    clean_auth_policy_cache,
-    encrypt_password,
-    get_access_model,
     is_anonymous,
 )
 from ._constants import PERMISSION_PRODUCT_LOGIN_KEY
@@ -20,13 +17,9 @@ from ._identity_web import forget_identity, remember_identity
 __all__: tuple[str, ...] = (
     "PERMISSION_PRODUCT_LOGIN_KEY",
     "AuthContextDict",
-    "check_password",
     "check_user_authorized",
     "check_user_permission",
-    "clean_auth_policy_cache",
-    "encrypt_password",
     "forget_identity",
-    "get_access_model",
     "is_anonymous",
     "remember_identity",
 )
